Completed the videoModeA fadeIn and fadeOut functions.
The fading functions are logic/draw blocking. The fade does not update the displayed tiles. Instead, VRAM is copied to the WebWorker that will draw a combined layer canvas. From that source data all the fade layers are created and then sent back to the client to be drawn in order.
Added _JSG.shared function "msToFrames". Will calculate how many complete frames are needed to make up the milliseconds requested.
Cleaned up the WebWorker handler in videoModeA.
WebWorker:
 - init, update_vram, and fade are working.
 - Added drawFromVRAM.
 - Finished fade.